LPS-87529 fix for doclib
